Fix concurrency issues when adding new data to models, also test the region online.
1. the main MAPE loop now can work with arbitary samlping interval, we fix
an issue where the pending value and the value could somehow overlap with
each others when many MAPE loop occur in parallell.

It now work by putting a sequencer on the ControlBus. However, currently it requires
each MAPE to be executed, we might be able to reduce the number of MAPE that needs to
be performed, e.g., by abort some MAPE loop before the last one and after
the next awaiting one.

2. test on regioning online, it now able to partition the region whenver there
is a significant model change.

ref: T7
type: consolidation/testing